Implement add_recordings and delete_recordings endpoints
Two endpoints and the relevant tests are implemented by this commit. They both make use of an already existent `update` method from db.dataset.
I have taken a design decision to use this method for several reasons:
- The fact it already exists, but is not yet used anywhere, suggests that it was created in a preparation for the forecoming development.
- The method already exists and ensures the correct structure of dataset with it's helper validation method.
- There is neither `class` nor `recording` module in the db lib and the task requires using methods from db.datasets. I believe that `recording` or `class`-specific methods should go to their own libs rather than to db.datasets.

Implementing these libs properly would take much more than 4 hours, but in a long term would be a better solution. Why? Because we could avoid recreating all classes and recordings every time we want to only add or delete a couple of recordings, which would be much more efficient regarding DB. However, in the current state of API, this doesn't seem to be a problem, as there are no endpoints for requesting single class or and classes belong exclusively to single dataset resource, so recreating them doesn't cause any internal app problems.

Also a DELETE `dataset` endpoint was fixed and it works now.